Module testnet_parachains_constants::westend::consensus

Consensus-related.

Constants

BLOCK_PROCESSING_VELOCITY

How many parachain blocks are processed by the relay chain per parent. Limits the number of blocks authored per slot.

MAXIMUM_BLOCK_WEIGHT

We allow for 2 seconds of compute with a 6 second average block.

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

RELAY_CHAIN_SLOT_DURATION_MILLIS

Relay chain slot duration, in milliseconds.

SLOT_DURATION

UNINCLUDED_SEGMENT_CAPACITY

Maximum number of blocks simultaneously accepted by the Runtime, not yet included into the relay chain.